backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.2
opencv-python==4.8.0.76
torch==2.0.1
torchvision==0.15.2
yolov5==7.0.13
tensorflow==2.13.0
librosa==0.10.0
google-cloud-speech==2.21.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
gunicorn==21.2.0
